Move masterdata to data.archive
